Producer/requirements.txt
pika
python-dotenv
python-logstash